- Pre-filter the constructors on argument types
- Cache when no default constructor is found (common for struct)

Benchmark was extracted from [here](https://github.com/dotnet/performance/blob/694c6bbe4e5b57e215a8ef689193cf8db0600ffb/src/benchmarks/micro/coreclr/System.Reflection/Activator.cs#L18-L22):

Before

|                      Method |     Mean |   Error |  StdDev |
|---------------------------- |---------:|--------:|--------:|
| CreateInstanceGenericStruct | 633.0 ns | 3.83 ns | 2.99 ns |
|  CreateInstanceGenericClass | 336.2 ns | 0.95 ns | 0.84 ns |

After

|                      Method |     Mean |   Error |  StdDev |
|---------------------------- |---------:|--------:|--------:|
| CreateInstanceGenericStruct | 274.8 ns | 0.62 ns | 0.55 ns |
|  CreateInstanceGenericClass | 335.9 ns | 0.61 ns | 0.51 ns |