@filipnavara 2019-11-02
I also tried to move the check around (https://github.com/filipnavara/mono/commit/c58ef2554d7b7d5608439ff689f3b8007c8973bf) and it results in some small additional speed up:

|                      Method |     Mean |   Error |  StdDev |
|---------------------------- |---------:|--------:|--------:|
| CreateInstanceGenericStruct | 243.1 ns | 1.89 ns | 1.77 ns |
|  CreateInstanceGenericClass | 299.7 ns | 5.43 ns | 4.81 ns |

However, I forgot to update the other caller of `GetDefaultConstructor`. I may revisit it later or just speed up the checks through other means.