@vargaz 2019-11-03
Since Volatile.Read/Write contains a barrier, I think this is ok now concurrency wise.